+++ examples/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workglow/cli
 
+## 0.0.79
+
+### Patch Changes
+
+- Merge and Split
+- Updated dependencies
+  - @workglow/tasks@0.0.79
+  - @workglow/ai@0.0.79
+  - @workglow/ai-provider@0.0.79
+  - @workglow/storage@0.0.79
+  - @workglow/task-graph@0.0.79
+  - @workglow/test@0.0.79
+  - @workglow/util@0.0.79
+
 ## 0.0.78
 
 ### Patch Changes