+++ packages/test/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @workglow/test
 
+## 0.0.79
+
+### Patch Changes
+
+- Merge and Split
+- Updated dependencies
+  - @workglow/tasks@0.0.79
+  - @workglow/ai@0.0.79
+  - @workglow/ai-provider@0.0.79
+  - @workglow/job-queue@0.0.79
+  - @workglow/sqlite@0.0.79
+  - @workglow/storage@0.0.79
+  - @workglow/task-graph@0.0.79
+  - @workglow/util@0.0.79
+
 ## 0.0.78
 
 ### Patch Changes