+++ packages/util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @workglow/util
 
+## 0.0.79
+
+### Patch Changes
+
+- Merge and Split
+
 ## 0.0.78
 
 ### Patch Changes